Re-scope container_puppet_config
The container_puppet_config module was used to run puppet to generate
config files and update config hash. However in EDPM we no longer use
puppet but use this module to update config hash.

This reduces scope of the module by removing the implementation to run
puppet because that part is no longer used.

Note that this change drops _exists method which is just a wrapper
of os.path.exists .

Signed-off-by: Takashi Kajinami <tkajinam@redhat.com>