Move SELinux task into edpm_bootstrap
Instead of having to interface with the task directly, move it into the
edpm_bootstrap role. The variable controlling the SELinux mode is also
renamed to edpm_bootstrap_selinux_mode so that it starts with the role
name.

As the deploy-edpm-selinux.yml playbook is no longer needed, it is
deleted.

Signed-off-by: James Slagle <jslagle@redhat.com>