Correct how molecule is launched
This patch bypasses the ci-framework "molecule_nodeps" target and
directly calls the script.

It also bind-mount the ci-framework repository elsewhere in the tree in
order to avoid potential issues with ansible auto-discovery
capabilities.

This patch was tested with a non-modified ci-framework repository, and
was successful with an existing container image as well as with a brand
new built image (from that very same Makefile).