remove remaining docs and update ovn and kernel roles
this patch removes the docs for the edpm nova compute,
libvirt and nvdimm roles.

This change removes the depencies on nova form the
ovn and kernel roles.

- kernel: the nova.conf will not exist when the kernel role is
executed and thsi corss depency was already volating the principal
that data should be passed in not looked up form the sideefects
of other roles. To address this the tsx check is removed.

- ovn: the ovn role did not have a direct depency on the nova role
but it was incorrectly using the nova and placement users.
all interaction form the ovn metadta agent shoudl use neutron
credentials. eventurally the config generation shoudl be removed
entirely from the ovn role as that shoudl be done by the neutron operator
so this patch just fixes the username and leave the rest for a later
pr to adress.