Default edpm_ceph_client_files_source to /etc/ceph
Remove task to fail role if edpm_ceph_client_files_source
is empty/undefined. Default edpm_ceph_client_files_source
and edpm_ceph_client_files_config_home to /etc/ceph.

If edpm_ceph_client_files_source does not exist on Ansible
host (e.g. stat returns "No such file or directory"), task
will warn ("Unable to find '/etc/ceph' in expected paths")
but not fail.

The dataplane-operator will invoke this role only when the
extraMounts has an extraVolType of Ceph. In that case it is
expected that /etc/ceph is mounted in the AEE pod as it is
for other pods which use Ceph (e.g. Cinder or Glance). This
role will then copy files in that directory to the EDPM node.

This change makes it easier for users since they don't need
to override the variables edpm_ceph_client_files_config_home
and edpm_ceph_client_files_source in the ansibleVars of the
OpenStackDataPlane CR.

Signed-off-by: John Fulton <fulton@redhat.com>