Get rid of tenacity
The ansible-core package in CentOS 8 Stream and CentOS Stream 9 will
use python runtimes which are different from the system default one.
Because of this change, ansible controller is no longer able to consume
the libraries installed in system default.

This removes the dependency on tenacity library installed in system
default python to avoid the breakage caused by that change.

Signed-off-by: Takashi Kajinami <tkajinam@redhat.com>