Molecule test skeleton for edpm_chrony role
The converge playbook is failing due to the permissions issue.
The issue was reported previously, and was resolved, in these cases,
however it persist in our situation.

Furthermore, allowing our container access to system clock is not
necessarily a good idea. Until the situation is resolved, this role
will only possess minimum test configuration, with a view that it will
be fixed at later date.

Signed-off-by: Jiri Podivin <jpodivin@redhat.com>