Molecule tests as github actions
Molecule tests for each role now have their own github job.
The list of roles to be tested is truncated for now, as not all roles have
working tests to execute.

Tests will be triggered by changes in one of the designated repositories.
Since all roles can be affected by changes in plugins, external dependencies and
github actions themselves, not to mention their possible mutual interactions,
it is better to run tests aggressively.

Signed-off-by: Jiri Podivin <jpodivin@gmail.com>