Remove edpm_nvdimm role as no longer supported
Intel discontinued supported for there intel optane datacenter memory
aka vPMEM aka persistent memory in august of last year.

As a result we decided not to graduate it form tech preview in 16.1 to
GA in 17.0 (thus the NextGen 18.0 as well) and instead remove support
for it in our product instead.

Also remove no longer used _pmem_ vars

Signed-off-by: Bohdan Dobrelia <bdobreli@redhat.com>