Replacing IMG_BASE_TAG with IMAGE_TAG_BASE in README.md
The guidelines offered by README.md for building and pushing an image
were referring to environment variable IMG_BASE_TAG, which is not
considered by the Makefile or the related scripts.

Following the guide would lead to unauthorized access exception upon push,
since the default value of the IMAGE_TAG_BASE would be used instead.

Signed-off-by: Jiri Podivin <jpodivin@gmail.com>