port deploy-libvirt playbook to edpm_libvirt role
This change ports the deploy-libvirt playbook form
the nova operator and converts it into a standalone
role.

This includes a molecule testing leveraging the
vagrant molecule plugin to create a vm for testing.
basic verifcation tests are also provided to ensure
the container are created, running and defiend as
systemd services and ensure the expected host
directories are created.