Add tags, update bindmoutnts and add zuul job
This change adds tags to the libvirt operator
and upates the bind mounts to add :shared,z
to /run/libvirt and /var/lib/nova

This change also refactors the molecule test_helpers
into the top level molecule dir and converts to useing the delegated
driver. The test-helpers will be reused in a future
edpm_nova role. As the molecule test now modifies the system
where it is run it is excluded from local testing.

The change is tested by a new zuul job which uses the molecule test
to verify the role.