Port nova playbook to edpm_nova role
This Change imports the nova-operator deploy-nova playbook
and converts it to a role.

The tasks have been broken down into a configure, install and as
yet unused post-install set of tasks.

The role assumes that config will be provided by mounting
configmaps into subdirectories of /var/lib/openstack/config
although this is overridable by setting edpm_nova_config_src
to an alternitive path. This is used in the molecule testing
to use static test-data inplace of operator generated configs.

As with the edpm_libvirt role the edpm_nova roles use
the delegated driver for testing and development
addtionally the edpm_nova molecule env uses the edpm_libvirt
role to provide libvirt in the test env.

This change uses this molecule testing to provide a zuul
job.