Add new network template
With the new CI layout, having a dedicated nodeset for CRC alongside a
dedicated Compute, we need to get a dedicated network template file in
order to consume the right interface - eth1, aka nic2.

That "nic2" interface is associated to a private, dedicated network in
order to allow network isolation to proceed, exactly like late OVB jobs
in TripleO CI ecosystem.

This patch therefore adds a new "ci" dedicated template we can then
consume from within the ci-framework, ensuring we're using the correct
interface.

Note: this is hopefully a temporary workaround - the ansibleVars we
should use is, for now, a string, and we can't edit it with kustomize.
There's a plan to change its type, allowing proper edition of the
mapping.

Signed-off-by: Cédric Jeanneret <cjeanner@redhat.com>