Replace edpm_network_config_template with existing _override
Given the way this interface is leveraged via the dataplane-operator, it
doesn't make sense to define a file path. Instead, this change moves to
using the existing edpm_network_config_override as the
edpm_network_config_template. This allows users to define jinja2 strings
for the variable, and they are rendered out to compute nodes using the
copy module.

Signed-off-by: Brendan Shephard <bshephar@redhat.com>