Replace edpm_network_config_template with existing _override
Given the way this interface is leveraged via the dataplane-operator, it
doesn't make sense to define a file path. Instead, this change moves the
existing edpm_network_config_override to be the primary interface. It
has subsequently been renamed to edpm_network_config_template.
This allows users to define jinja2 strings for the variable, and they are
rendered out to compute nodes using the copy module.

Depends-On: https://github.com/openstack-k8s-operators/dataplane-operator/pull/318
Signed-off-by: Brendan Shephard <bshephar@redhat.com>