Fix docs job for local execution
this change fixes the doc job for the libvirt role
and also correct an error where the docs/source repo did not have
the expected permissions. finally the docs-requirements.txt
was updated and build-docs.sh updated to use upperconstraint
this prevent pip backtracking for 10s of minutes trying
to resolve the package requirements.