remove O(n*m) loop from edpm_kernel
the orginal commit of edpm_kernel used the ansible.posix.sysctl modules
to set extra sysctl options in a loop. later pr #157
added the extra sysctl arg to the template and too the loop.

By adding edpm_kernel_sysctl_extra_settings to the template, pr #157
removed the need for the loop and usage of the ansible.posix.sysctl
module entirely. This change simply removes it and relys on the
exisitng systemd service restart to pick up the cahnge when the file
is updated.

Closes: #254