Use ansible_nodename for ovn chassis hostname
It was originally changed from ansible_fqdn
to ansible_hostname as part of [1] but that caused
issues in environments where a fqdn is set as hostname.

We need to keep the ovn chassis hostname in sync with how
nova-compute "host" config is set.
Nova "host" defaults to socket.gethostname() and to
sync with it we need to use ansible_nodename as
ansible_hostname strips out domain part.

[1] https://github.com/openstack-k8s-operators/edpm-ansible/pull/211

(cherry picked from commit 5a3559b5da5882e49b2dd00201a90115facc519c)