edpm_libvirt should reload NFT ruleset after configuring it
The post-install tasks file was calling the configure tasks
in the edpm_nftables role to update the rules file in the
/etc/nftables/ directory, but was not calling the run tasks
to make the firewall reload the ruleset. This caused the
output of `nft list ruleset` to not include the VNC rule
which the edpm_libivirt role adds. Fix this by calling
the run tasks file.

Update molecule tests for edpm_libvirt and edpm_nova to
open port 22 in firewall. Otherwise the new firewall reload
results in port 22 being blocked during log collection which
causes the molecule tests to fail.

Jira: OSP-27573
Signed-off-by: John Fulton <fulton@redhat.com>