Don't override the compute routing
This patch will ensure the compute is still reachable in the CI
ecosystem, since the default route should be on the public interface and
not touched.

Note: once we're able to switch to the multinode layout, we will refine
the job and leverage the proper networkConfig.template, allowing to
remove this file. But we can't run too many rabbits at the same time.