Add ceph_rgw and ceph_rgw_frontend in te list of enabled_services
When HCI is deployed and the Ceph firewall rules are applied, we
need to make sure that if RGW is deployed in the orchestrator both
backends and the related frontend are able to communicate. RGW can
be deployed at day1 or as part of day2 operations, and eventually
it can be updated to use tls.
By adding radosgw among the enabled_services by default we can make
sure to not hit firewall related issues.
In CI we use the Storage network to reserve a VIP for this service,
however, in a production environment an administrator can choose to
expose RGW using a public network. For this reason both 'ceph_rgw'
and 'ceph_rgw_frontend' are added to the default list.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>