add support for libvirt_guests to edpm_libvirt
This change ports the libvirt_guests code from edpm_nova_libvirt
to edpm_libvirt and adapts it to follow the conventions of the
new role including molecule testing.

libvirt guests is a script that ensure that vms are gracefully shutdown
when the host is requested to reboot or poweroff.
This service is not strictly required however it can reduce the risk
of guest filesystem curruption. Wihtout this service systemd
will still attempt to terminate the qemu instance with a much
shorter timeout before sending sig_kill.

As there is little reason not to always use libvirt_guests
to protect the guest filesystem this change makes its use
mandataory and does not allow it to be disabled unlike
the previous opt in behavior.

As part of this refactoring the verify_systemd_service.yaml
test helper has been renamed to verify_systemd_unit.yaml and
extened to make asserting the enabled and runnign state
configurable. the edpm_nova role has also been adapted to
that change.

Closes: OSPRH-156