extract host user,group and dir creation to edpm_users role
This change factors out the code to create users/groups that must
map to the same uid/gid within or outside of a container.

This is required to supprot ovs-dpdk where the hugepage and
vhost user socket permission must align to enabel ovs-dpdk
running on the host to comunciate with qemu running in a container.

Thsi work enabled the future removal of the edpm_nova_compute
and edpm_nova_libvirt roles.