edpm_libvirt role should handle quoted cephx keys
It's valid for the key inside a ceph keyring file to be
either quoted or not quoted. Before this patch, if the
key was quoted, then it was not passed to the "virsh
secret-set-value" command because it did not satisfy the
regex check because of the quotes. This change strips the
quotes from the key value before it is checked and passed
to the virsh command.

Signed-off-by: John Fulton <fulton@redhat.com>