Revert "Do not attempt to manage collocated Ceph containers"
This reverts commit 0e6058cc47c5f147a2338b41f07d6f8dd8bac5de.

This breaks HCI but HCI is not yet in CI so a follow up patch
will fix what this patch breaks for people deploying HCI.

The commit being reverted introduced a task that we still
the functionality of, however, its implementation was not
efficient and takes too long to run in the CI as documented:

https://github.com/openstack-k8s-operators/data-plane-adoption/pull/171

We still need to exclude the ceph containers but will need
to do it with less computational resources (e.g. maybe pure
jinja will be faster or we might need a custom filter).

Conflicts:
	roles/edpm_container_manage/tasks/delete_orphan.yml