Add bootstrap task to the neutron_dhcp role
This bootstrap will make sure that openvswitch is running on the
host before dhcp agent's container will be started.
Previously I added it by mistake to the molecule's prepare module as it
was causing failing molecule test when openvswitch wasn't running on the
fake_host but it shouldn't be done only in test but always.