[edpm_ssh_known_hosts]Support multiple host key types
Originally the role only copied the RSA host keys to the generated
known_hosts file. However newer ssh version prefer ED25519 over RSA.
This resulted that even though the known_hosts file was present in the
EDPM nodes when an SSH connection is initiated between the nodes the
user is got prompted to accept the host key. This is against the purpose
of the role.

So this PR adds support for ED25519 and ECDSA host keys as well while
keep supporting RSA as well.