ovn: actually apply hw-offload setting
The code was constructing a new, separate dict but never applying it.

A new molecule scenario is introduced for hwoffload case, and test setup
from the `default` scenario is reused (by setting up symlinks where
appropriate as well as moving common verify tasks into a separate tasks
file that is included in both `default` and `hwoffload` scenarios.)

Note: `noconfig` scenario should also be migrated to symlinks to avoid
code duplication, but this exercise is left for a separate patch.

Signed-off-by: Ihar Hrachyshka <ihrachys@redhat.com>