[nova]Set up ssh keys auth for the nova user
The nova user needs to be able to copy files between the compute nodes.
The edpm_nova role now expects that a Secret is mounted to the AnsibleEE
CR containing an ssh keypair (ssh-privatekey and ssh-publickey). This
keypair is used to set up the auth for the nova user between the
compute nodes. On the EDPM host the nova user will have the public key
added as authorized_keys. While the nova user in the nova-compute
container will have the private key added as identity in
/var/lib/nova/.ssh. Also an ssh config file is added in the container to
ensure that this identity is used for all the outgoing ssh connection.