[molecule]Fix known_hosts molecule test
By merging #448 after #449 caused that the molecule test is now failing
on main. Separately the two PRs wer doing the correct thing. The github
action doing the pre-merge testing does not do a rebase before running
the test so it cannot detect if the merged in PR fails the test.

This PR fixing the molecule test.

[1] https://github.com/openstack-k8s-operators/edpm-ansible/pull/448/files#diff-654bc2164f880d6ff16c6009ecd64121cac7d59497ada186aab1ffded0a466b8