add optional vagrant senairo for local testing and development.
The libvirt and nova roles currently uses the delegated driver
which runs molecule agaisnt localhost. This is requried for
ci as the podman driver cannot be used with this role however
it also means you must use a seperate development vm.

This change readds supprot for using the molecule vagrant plugin
for this role and configures the vagrant senario to use the
prepare, converge and verify playbooks from the default delegated
senario.

Addtionally this change updates yamllint and ansible-lint to not
scan files in a venv i.e. .venv in the root of the edpm-ansible
repo. .venv is already ignored by .gitignore so this also only
affect local development.

the molecule-plugins min version is raised to 23.5.0 to include
https://github.com/ansible-community/molecule-plugins/pull/142