Refactor libvirt to be host installed
This commit removed installation of of libvirt as a
container and converts to a host level install.

The molecule tests are enhanced to allow testing this and
we revert to using ceph config from /etc/ceph.

The nova roles is also updated to no longer relable /run/libvirt
and the prepare palybook is updated to run repo_setup to ensure
ceph-common can be isntalled.