[edpm_nova]Configure my_ip to ctlplane_ip for nova-compute
This is needed as nova-compute, if my_ip is not configured,
tries to figure out its own IP address in complicated way and in a
system with multiple network interfaces there is no guarantee that
it selects the IP of the designated control plane interface.

The my_ip configuration is EDPM node specific so it cannot be generated
globally for a cell in nova-operator. So this is generated by ansible
now.

There is an upstream nova blueprint to eventually get rid of the need of
this configuration:
https://blueprints.launchpad.net/nova/+spec/libvirt-migrate-with-hostname-instead-of-ip