Add support for the sidecar containers in neutron_dhcp role
Similarly to what was done in Tripleo, neutron agents which are spawning
some long running processes, like e.g. haproxy or dnsmasq should run them
in the sidecar containers to avoid breakege in the dataplane when
e.g. agent's container is restarted.

This patch adds wrappers for haproxy and dnsmasq sidecars in the
neutron_dhcp role.
Those wrappers script are taken almost directly from the TripleO's
puppet module [1] and are just converted to jinja2 format and to support
only Podman as container_cli.

This patch adds also support for kill-scripts which is used to stop
and clean sidecar containers. It is takend almost directly from the
TripleO Heat Templates [2].

[1] https://opendev.org/openstack/puppet-tripleo/src/branch/master/templates/neutron
[1] https://github.com/openstack/tripleo-heat-templates/blob/master/deployment/neutron/kill-script