Fix molecule `default` scenario for `edpm_network_config` role
Since molecule tests will be executed in a container, we can't configure an ovs bridge on top of a dummy interface or a tap interface.
The scope of this test isn't to provide a "working" edpm scenario example, so we are going to provide a sample os-net-config configuration which works in a container on top of a dummy interface.

Signed-off-by: Roberto Alfieri <ralfieri@redhat.com>