Mount TLS certs
The dataplane operator will provide certs and keys to the ansibleEE pod
in /var/lib/openstack/certs as well as a cacert file in
/var/lib/openstack/cacerts.  It will also pass an ansible variable
tls_certs_enabled to indicate that the certs are present.

This PR copies these files to the appropriate place on the compute node
and then mounts these files in the nova-compute container.

We still need to confirm that these files are in the right location
in the container and add config to use these certs - as well as code
to trust the cacert.  Will need help from compute DFG for that.

I'm removing the dependency on the dataplane operator patch that adds
these files, because it isn't a strict dependency.  An ansible
variable must be set to trigger this new code.