Execute `edpm_bootstrap_command` during `download_cache`
Since the registration of the EL nodes is performed during the `edpm_bootstrap_command`
task in the `edpm_bootstrap` role, we moved that task into a new task
file called `bootstrap_command.yml` so it can be called by the
`edpm_download_cache` role without invoking the entire role. We also
added a marker that will avoid to execute the task twice if it's been
already executed.

Signed-off-by: Roberto Alfieri <ralfieri@redhat.com>