Set some neutron-ovn-metadata agent options only if necessary
Some metadata agent config options are usually provided by the nova and
ovn operators and are set in separate config files.
To avoid confusion with empty values in the main config file used by the
neutron-ovn-metadata-agent, those options will be set there only if they
are provided in the role's parameters.