This commit adds a new edpm_ovs role to manage the installation of openvswitch. As part of this change, the installation is moved from the bootstrap playbook to the configure_network playbook. To aide with testing the verify_sockets_exists test helper is moved from the edpm_libvirt role to the top-level molecule test-helpers and refactored to work with any unix socket. The edpm_ovs role uses this to validate that the ovs-db socket is created with the expected permission.
As part of this refactor a latent bug where the OVS user was not
created before the OVS package was installed has been fixed.
This will ensure that the OVS user on the host and in the containers
are aligned to the static user/group ID used in the container images.

Finally, a new zuul job is added to run the edpm_ovs tests.
Closes: OSPRH-3451