refactor edpm_neutron_sriov role
This change replaces the use of edpm_container_standalone
with edpm_container_manage. edpm_container_standalone internally
uses edpm_container_manage. To enable that the yaml files
are converted to json format as edpm_container_manage requires json

This also refactors the use of become to ensure that only the tasks
that require elevated permissions are elevated. In a similar
vain the permession of the tls certs are restricted to 0640 as they
should not be world readable.

to align with nova, ceph, tls and other roles, the config dir on the
edpm host used by the neutron sriov nic agent has been updated
to /var/lib/openstack/config/neutron-sriov-agent
from /var/lib/config-data/ansible-generated/neutron-sriov-agent

Finally debug logging is enabeld by default.
We want to ship with debug logs on by default to enabled
debuging using collected logs in a production envionment without
needing to recreate a failure.