Add rdoproject repo to update libvirt
There is a bug on libvirt 9.10 and libvirt 10 which fixes this bug is
not yet available. In order to workaround this situation, the rdoproject
build latest libvirt 10 version on their repo and we are adding it here
in order to fix our long standing block on edpm jobs.
I used shell to configure the repo, because I wasn't able to find in the
ansible documentation how to do so with a .repo file.