Added FIPS related tasks in `edpm_bootstrap` and `edpm_iscsid` roles
These tasks can be used to check the status, enable or disable FIPS on a
node.

They rely on `fips-mode-setup` tool provided by the `crypto-policies-scripts`
package.

The action is governed by the `edpm_fips_mode` variable which can
assumes 3 values:

- `enable`, to enable FIPS
- `disable`, to disable FIPS
- `check`, to check the status of FIPS

It defaults to `check`.

In order to test the FIPS status, the `edpm_bootstrap` molecule
tests are now performed on a vm rather than a container.

Closes: https://issues.redhat.com/browse/OSPRH-3071

Signed-off-by: Roberto Alfieri <ralfieri@redhat.com>