[edpm_nova]Use FQDN as sevice hostname
The greenfield deployment sets up host FQDN via IPAM but did not change
the /etc/hostname to FQDN. This causes that libvirt will see and use
FQDN for the node names, but nova-compute uses the sort hostname
returned by socket.gethostname().

In greenfield it causes that the nova-compute host name is different
form the compute nodename. Nova works with this setup, but it will cause
confusion for the users coming from older versions where both names are
FQDN.

During adoption, when the adopted cloud does not have FQDN in the
/etc/hostname, the compute service host name will change as in 18 we are
not yet hardcoding the FQDN in the nova.conf. This results in failed
adoption.

This commit ensures the [DEFAULT]host is set to the FQDN based on the
canonical_hostname variable.

This patch does not handle the adoption case where:
* the old cloud had something other than the FQDN set in the service
  host configuration
* the old cloud had a different FQDN set than what is set in the
canonical_host in the new cloud configuration

There will be a separate patch to solve the adoption case via a new pre
adoption validation role.

Fixes: OSPRH-5197