use canonical_hostname not inventory hostname for certs
the tls certs are indexed by the node.HostName
when generating the invetory the inventory_hostname is
computed strings.Split(node.HostName, ".")[0]
node.HostName is stored in the canonical_hostname field
of the host vars.

This change updates edpm_install_certs to align
to the go code by using canonical_hostname instead.