[edpm_nova]Do not add immutable flag to compute_id
The podman version 4.6.0 has a bug that causes that sometimes a
container fails to start if it mounts an file with +i flag set.
This podman version is used in the adoption form state making the
adoption unstable (1 in 8 CI run fails). The greenfield 18 deployment
uses podman 4.9.4 that does not show this error. So to make the adoption
stable the edpm_nova role will not add the +i flag to existing
compute_id files. Still if the role pre-generates the file that means it
is a greenfield run so the flag can be, and will be, set safely.