Check an FQDN is set
An FQDN should be set, and a validation is added to check. The
validation can be disabled with the edpm_nodes_validation_check_for_fqdn
variable.

The molecule prepare.yml is updated to set a hostname so that the check
passes.

The existing validation that validates the FQDN is set in /etc/hosts is
updated to use a new variable,
edpm_nodes_validation_validate_fqdn_hosts_file, for the location of the
hosts file so that the molecule test can still run and pass using a test
hosts file at /tmp/hosts.

Jira: https://issues.redhat.com/browse/OSPRH-6187
Jira: https://issues.redhat.com/browse/OSPRH-16536
Signed-off-by: James Slagle <jslagle@redhat.com>