Fix pre adoption nova.conf path default
The 91d657948bcfa5159e02f4ae4dae579c58a7bfbc wrongly used
/var/lib/config-data/nova instead of nova_libvirt to look at the config
of from of the pre-adopted node. We missed this as it was only tested
with standalone deployments where both nova and nova_libvirt exists on
the machine. In a multinode deployment only nova_libvirt exists on the
compute node.

Implements: OSPRH-6333