[pre-adoption-validation]Check for kernel args changes
The edpm_kernel role checks is the content of the edpm_kernel_args
is already in the /proc/cmdline and if it isn't then it reconfigure grub
and request a reboot. This pre-adoption check therefore checks the same
and fails the adoption early to avoid kernel arg reconfiguration during
the edpm_kernel role run.

Implements: OSPRH-5714