When kernelargs are changed, ensure reboot file exists
This change ensures that the `edpm_kernel` file is created under the
`reboot_required` directory when `grub.cfg` has been updated. This
ensures the user is able to determine when a node needs rebooting[1], and
they can do so by including our `edpm_reboot` role via the `reboot.yml`
playbook in a `servicesOverride` on the deployment.

[1] https://github.com/openstack-k8s-operators/edpm-ansible/blob/main/docs/source/roles/role-edpm_reboot.rst?plain=1#L10-L18

Signed-off-by: Brendan Shephard <bshephar@redhat.com>