Fix grub2-mkconfig kernel args for RHEL 9 and up
In order to update the kernel args properly in RHEL 9.X the --update-bls-cmdline is required.Without this flag the /boot/loader/entries/ files are not updated which causes the kernal args in the grub.cfg to be ignored. Added the flag for RHEL9 and up to allow proper cmdline update during boot.

This patch should fix this issue: https://issues.redhat.com/browse/OSPRH-6665