[edpm_pre_adoption]Ensure that hostname is consistent
Adds a pre-adoption check to ensure that the nova-compute service
hostname is consistent what hostname -f return and therefore what
libvirt reports as hostname. We need to ensure this as in the 18
deployment we enforcing a single hostname via canonical_hostname to both
nova and libvirt. So during adoption one of the hostname would be
changed that would cause nova-compute issues.

Related: OSPRH-6335