FIPS task should not change reboot policy
As reported in https://issues.redhat.com/browse/OSPRH-7301 , the
fips task incorrectly changes the reboot policy when the required
fips status changes.  We should follow whatever is defined by the
deployer instead.

If nothing is defined, the policy is auto - which will result in
a reboot.

In the case of a redeployed or adopted environment, we will end up
simply logging that a reboot is supposed to happen.  But we will
then likely fail the subsquent fips status check - which is the
desired behavior.